palette and cleanup
